// src/types/ProductTypes.ts
export type ProductTypes = {
    id: number;
    title: string;
    price: number;
    description: string;
    category: string;
    image: string;
    rating: {
        rate: number;
        count: number;
    };
};

//Esta es la seccion de tipados donde especificamos los elementos de la api o json